+++ src/Phan/Analysis/ConditionVisitor.php
@@ -569,15 +569,36 @@ class ConditionVisitor extends KindVisitorImplementation
     private static function initTypeModifyingClosuresForVisitCall() : array
     {
         $make_basic_assertion_callback = static function (string $union_type_string) : Closure {
-            $type = UnionType::fromFullyQualifiedString(
+            $asserted_union_type = UnionType::fromFullyQualifiedString(
                 $union_type_string
             );
+            $asserted_union_type_set = $asserted_union_type->getTypeSet();
+            $empty_type = UnionType::empty();
 
             /** @return void */
-            return static function (Variable $variable, array $args) use ($type) {
+            return static function (Variable $variable, array $args) use ($asserted_union_type, $asserted_union_type_set, $empty_type) {
+                $new_types = $empty_type;
+                foreach ($variable->getUnionType()->getTypeSet() as $type) {
+                    $type = $type->withIsNullable(false);
+                    if ($type->canCastToAnyTypeInSet($asserted_union_type_set)) {
+                        $new_types = $new_types->withType($type);
+                    }
+                }
+
                 // Otherwise, overwrite the type for any simple
                 // primitive types.
-                $variable->setUnionType($type);
+                $variable->setUnionType($new_types->isEmpty() ? $asserted_union_type : $new_types);
+            };
+        };
+        $make_direct_assertion_callback = static function (string $union_type_string) : Closure {
+            $asserted_union_type = UnionType::fromFullyQualifiedString(
+                $union_type_string
+            );
+            /** @return void */
+            return static function (Variable $variable, array $args) use ($asserted_union_type) {
+                // Otherwise, overwrite the type for any simple
+                // primitive types.
+                $variable->setUnionType($asserted_union_type);
             };
         };
 
@@ -657,9 +678,10 @@ class ConditionVisitor extends KindVisitorImplementation
             $variable->setUnionType($new_type);
         };
 
-        $float_callback = $make_basic_assertion_callback('float');
+        // Note: LiteralIntType exists, but LiteralFloatType doesn't, which is why these are different.
         $int_callback = $make_basic_assertion_callback('int');
-        $null_callback = $make_basic_assertion_callback('null');
+        $float_callback = $make_direct_assertion_callback('float');
+        $null_callback = $make_direct_assertion_callback('null');
         // Note: isset() is handled in visitIsset()
 
         return [
@@ -677,7 +699,7 @@ class ConditionVisitor extends KindVisitorImplementation
             'is_numeric' => $make_basic_assertion_callback('string|int|float'),
             'is_object' => $object_callback,
             'is_real' => $float_callback,
-            'is_resource' => $make_basic_assertion_callback('resource'),
+            'is_resource' => $make_direct_assertion_callback('resource'),
             'is_scalar' => $scalar_callback,
             'is_string' => $make_basic_assertion_callback('string'),
             'empty' => $null_callback,