+++ tests/files/src/0516_literal_type_narrowing.php
@@ -1,0 +1,11 @@
+<?php
+call_user_func(function() {
+    $x = 2;
+    assert(is_int($x));
+    assert(is_numeric($x));
+    echo strlen($x);
+
+    $y = 'a string';
+    assert(is_string($y));
+    echo intdiv($y, 2);
+});